S9/requirements.txt
torch
torchvision
gradio
Pillow
requests
gdown 